requirements.txt
Flask
google-generativeai
python-dotenv
requests
gunicorn
Flask
mysql-connector-python
flask-bcrypt
google-generativeai
python-dotenv
requests
gunicorn
